Move random invoice number generator to Sylius\Component\Core\Payment
